+++ bin/config/patterns/config.patterns.ts
@@ -4,20 +4,28 @@ export const patterns = {
 const config: StestConfig = {
   pattern: 'test/**/*.{spec,test}.ts',
   ignore: ['node_modules'],
+  autoClearMocks: true,
+  cacheWatcher: false,
 };
 export default config;
 `,
   js: `const config = {
   pattern: 'test/**/*.{spec,test}.ts',
   ignore: ['node_modules'],
+  autoClearMocks: true,
+  cacheWatcher: false,
 };
 module.exports = config;
 `,
   json: `{
   "pattern": "test/**/*.{spec,test}.ts",
-  "ignore": ["node_modules"]
+  "ignore": ["node_modules"],
+  "autoClearMocks": true,
+  "cacheWatcher": false,
 }`,
   yml: `pattern: "test/**/*.{spec,test}.ts"
 ignore:
-  - node_modules`,
+  - node_modules
+autoClearMocks: true,
+cacheWatcher: false,`,
 };